+++ packages/serverless-workflow-service-catalog/src/channel/parsers/impl/convertSource.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2023 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { SwfCatalogSourceType, SwfServiceCatalogFunctionSource, SwfServiceCatalogServiceSource } from "../../api";
+import { SwfCatalogSourceType, SwfServiceCatalogFunctionSource, SwfServiceCatalogServiceSource } from "../../../api";
 
 export function convertSource(catalogSource: SwfServiceCatalogServiceSource): SwfServiceCatalogFunctionSource {
   if (catalogSource?.type === SwfCatalogSourceType.LOCAL_FS) {
