+++ packages/serverless-workflow-service-catalog/src/channel/parsers/impl/openapi/openapi.ts
@@ -22,104 +22,118 @@ import {
   SwfServiceCatalogService,
   SwfServiceCatalogServiceSource,
   SwfServiceCatalogServiceType,
-} from "../../api";
-import { convertSource } from "./ConvertSource";
-import * as AsyncApi from "../../api/asyncapitypes";
+} from "../../../../api";
+import { OpenAPIV3 } from "openapi-types";
+import { get } from "lodash";
+import { convertSource } from "../convertSource";
 
-export function parseAsyncApi(
+const APPLICATION_JSON = "application/json";
+
+type OpenapiPathOperations = Pick<
+  OpenAPIV3.PathItemObject,
+  "get" | "put" | "post" | "delete" | "options" | "head" | "patch" | "trace"
+>;
+
+export function parseOpenApi(
   args: {
     source: SwfServiceCatalogServiceSource;
     serviceFileName: string;
     serviceFileContent: string;
   },
-  serviceAsyncApiDocument: AsyncApi.AsyncAPIDocument
+  serviceOpenApiDocument: OpenAPIV3.Document
 ): SwfServiceCatalogService {
-  const swfServiceCatalogFunctions = extractFunctions(serviceAsyncApiDocument, convertSource(args.source));
-
+  const swfServiceCatalogFunctions = extractFunctions(serviceOpenApiDocument, convertSource(args.source));
   return {
-    name: serviceAsyncApiDocument.info.title ?? args.serviceFileName,
-    type: SwfServiceCatalogServiceType.asyncapi,
+    name: serviceOpenApiDocument.info.title ?? args.serviceFileName,
+    type: SwfServiceCatalogServiceType.rest,
     source: args.source,
     functions: swfServiceCatalogFunctions,
     rawContent: args.serviceFileContent,
-  } as SwfServiceCatalogService;
+  };
 }
 
 export function extractFunctions(
-  serviceAsyncApiDocument: AsyncApi.AsyncAPIDocument,
+  serviceOpenApiDocument: OpenAPIV3.Document,
   source: SwfServiceCatalogFunctionSource
 ): SwfServiceCatalogFunction[] {
-  const swfServiceCatalogFunctions = Object.entries(serviceAsyncApiDocument.channels).map(
-    ([endpoint, channelItem]: [string, AsyncApi.ChannelsObject]) => {
-      return extractChannelItemFunctions(channelItem, endpoint, serviceAsyncApiDocument, source);
+  const swfServiceCatalogFunctions = Object.entries(serviceOpenApiDocument.paths).map(
+    ([endpoint, pathItem]: [string, OpenAPIV3.PathItemObject]) => {
+      return extractPathItemFunctions(pathItem, endpoint, serviceOpenApiDocument, source);
     }
   );
-
   return [].concat.apply([], swfServiceCatalogFunctions);
 }
 
-function extractChannelItemFunctions(
-  channelItem: AsyncApi.ChannelsObject,
+function extractPathItemFunctions(
+  pathItem: OpenapiPathOperations,
   endpoint: string,
-  serviceAsyncApiDocument: AsyncApi.AsyncAPIDocument,
+  serviceOpenApiDocument: OpenAPIV3.Document,
   source: SwfServiceCatalogFunctionSource
 ): SwfServiceCatalogFunction[] {
   const swfServiceCatalogFunctions: SwfServiceCatalogFunction[] = [];
 
-  const functionArguments: Record<string, SwfServiceCatalogFunctionArgumentType> = {};
+  Object.values(pathItem)
+    .filter((pathOperation) => pathOperation.operationId)
+    .forEach((pathOperation: OpenAPIV3.OperationObject) => {
+      const body = pathOperation.requestBody as OpenAPIV3.RequestBodyObject;
 
-  // Looking at operation params
-  if (channelItem.parameters) {
-    extractFunctionArgumentsFromParams(channelItem?.parameters, functionArguments);
-  }
-  Object.values(channelItem)
-    .filter((channelOperation: AsyncApi.OperationObject) => channelOperation.operationId)
-    .forEach((channelOperation: any) => {
-      const body = channelOperation.message;
-      const name = channelOperation.operationId as string;
+      const name = pathOperation.operationId as string;
+
+      const functionArguments: Record<string, SwfServiceCatalogFunctionArgumentType> = {};
+
+      // Looking at operation params
+      if (pathOperation.parameters) {
+        extractFunctionArgumentsFromParams(pathOperation.parameters, functionArguments);
+      }
 
       // Looking only at application/json mime types, we might consider others.
-      if (body) {
-        extractFunctionArgumentsFromRequestBody(body ?? {}, serviceAsyncApiDocument, functionArguments);
+      if (body && body.content && body.content[APPLICATION_JSON] && body.content[APPLICATION_JSON].schema) {
+        extractFunctionArgumentsFromRequestBody(
+          body.content[APPLICATION_JSON].schema ?? {},
+          serviceOpenApiDocument,
+          functionArguments
+        );
       }
 
       const swfServiceCatalogFunction: SwfServiceCatalogFunction = {
         source,
         name,
-        type: SwfServiceCatalogFunctionType.asyncapi,
+        type: SwfServiceCatalogFunctionType.rest,
         arguments: functionArguments,
       };
       swfServiceCatalogFunctions.push(swfServiceCatalogFunction);
     });
+
   return swfServiceCatalogFunctions;
 }
 
 function extractFunctionArgumentsFromParams(
-  channelParams: any,
+  pathParams: (OpenAPIV3.ReferenceObject | OpenAPIV3.ParameterObject)[],
   functionParams: Record<string, SwfServiceCatalogFunctionArgumentType>
 ) {
-  const paramNames = Object.keys(channelParams);
-  paramNames.forEach((paramName) => {
-    if (channelParams && paramName) {
-      functionParams[paramName] = resolveArgumentType(channelParams[paramName]?.schema?.type);
+  pathParams.forEach((pathParam) => {
+    const name = get(pathParam, "name");
+    const type = get(pathParam, "schema.type");
+    if (name && type) {
+      functionParams[name] = resolveArgumentType(type);
     }
   });
 }
 
 function extractFunctionArgumentsFromRequestBody(
-  message: AsyncApi.MessageObject,
-  doc: AsyncApi.AsyncAPIDocument,
+  schema: OpenAPIV3.ReferenceObject | OpenAPIV3.SchemaObject,
+  doc: OpenAPIV3.Document,
   functionParams: Record<string, SwfServiceCatalogFunctionArgumentType>
 ) {
-  const schemaObject: AsyncApi.SchemaObject = extractMessageObject(message, doc);
+  const schemaObject: OpenAPIV3.SchemaObject = extractSchemaObject(schema, doc);
   if (schemaObject.properties) {
     Object.entries(schemaObject.properties).forEach(
-      ([propertyName, propertySchema]: [string, AsyncApi.ReferenceObject | AsyncApi.SchemaObject]) => {
-        const asReference = propertySchema as AsyncApi.ReferenceObject;
+      ([propertyName, propertySchema]: [string, OpenAPIV3.ReferenceObject | OpenAPIV3.SchemaObject]) => {
+        const asReference = propertySchema as OpenAPIV3.ReferenceObject;
         if (asReference.$ref) {
           functionParams[propertyName] = SwfServiceCatalogFunctionArgumentType.object;
         } else {
-          const asSchema = propertySchema as AsyncApi.SchemaObject;
+          const asSchema = propertySchema as OpenAPIV3.SchemaObject;
           if (asSchema.type) {
             functionParams[propertyName] = resolveArgumentType(asSchema.type);
           }
@@ -129,27 +143,17 @@ function extractFunctionArgumentsFromRequestBody(
   }
 }
 
-function extractMessageObject(message: AsyncApi.MessageObject, doc: AsyncApi.AsyncAPIDocument): any {
-  const asReference = message as AsyncApi.MessageObject;
-  if (asReference.$ref) {
-    const messageRef: any = asReference.$ref.split("/").pop() ?? "";
-    const resolvedMessage = doc.components?.messages
-      ? doc.components?.messages[messageRef]
-      : ({} as AsyncApi.MessageObject);
-
-    return extractSchemaObject(resolvedMessage?.payload, doc);
-  }
-  return message?.payload;
-}
-
-function extractSchemaObject(payload: any, doc: AsyncApi.AsyncAPIDocument): AsyncApi.SchemaObject {
-  const asReference = payload as any;
+function extractSchemaObject(
+  schema: OpenAPIV3.ReferenceObject | OpenAPIV3.SchemaObject,
+  doc: OpenAPIV3.Document
+): OpenAPIV3.SchemaObject {
+  const asReference = schema as OpenAPIV3.ReferenceObject;
   if (asReference.$ref) {
     const schemaRef = asReference.$ref.split("/").pop() ?? "";
     const resolvedSchema = doc.components?.schemas ? doc.components?.schemas[schemaRef] : {};
     return extractSchemaObject(resolvedSchema, doc);
   }
-  return asReference as any;
+  return schema as OpenAPIV3.SchemaObject;
 }
 
 function resolveArgumentType(type: string): SwfServiceCatalogFunctionArgumentType {
